menu refactoring. basic links implemented.
