+++ lib/editor/components/Editor.js
@@ -2,28 +2,32 @@ import React, { Component } from 'react'
 import { Editor as SlateEditor } from 'slate'
 import {
   Button,
-  NarrowContainer
+  Container,
+  NarrowContainer,
+  R,
+  Interaction,
+  colors
 } from '@project-r/styleguide'
 import Schema from '../schema'
-import createMarkToggleMenu from './createMarkToggleMenu'
-import createHoverToolbar from './createHoverToolbar'
-import createToolbar from './createToolbar'
+import MarkMenu from './MarkMenu'
+import BlockMenu from './BlockMenu'
+import LinkMenu from './LinkMenu'
+import HoverToolbar from './HoverToolbar'
+import SideToolbar from './SideToolbar'
 
 import FormatBold from 'react-icons/lib/md/format-bold'
 import FormatItalic from 'react-icons/lib/md/format-italic'
 import FormatUnderlined from 'react-icons/lib/md/format-underlined'
 import FormatStrikethrough from 'react-icons/lib/md/format-strikethrough'
 
-const MarkToggleMenu = createMarkToggleMenu([
+const marks = [
   ['bold', <FormatBold />],
   ['italic', <FormatItalic />],
   ['underline', <FormatUnderlined />],
   ['strikethrough', <FormatStrikethrough />]
-])
+]
 
-const HoverToolbar = createHoverToolbar([MarkToggleMenu])
-
-const Toolbar = createToolbar([
+const blocks = [
   ['p', 'p'],
   ['h2', 'h2'],
   ['lead', 'lead'],
@@ -31,7 +35,7 @@ const Toolbar = createToolbar([
   ['label', 'label'],
   ['blockquote', 'blockquote'],
   ['interaction.p', 'interaction.p']
-])
+]
 
 export default class Editor extends Component {
   constructor(props) {
@@ -39,6 +43,7 @@ export default class Editor extends Component {
     this.state = {
       state: props.state
     }
+    this.onChange = this.onChange.bind(this)
   }
 
   onChange(state) {
@@ -52,29 +57,56 @@ export default class Editor extends Component {
   }
 
   render() {
+    const { state } = this.state
     return (
       <div>
-        <Button
-          primary
-          onClick={() =>
-            this.props.onCommit(this.state.state)}
+        <Container
+          style={{
+            display: 'flex',
+            flexFlow: 'row no-wrap',
+            justifyContent: 'space-between',
+            borderBottom: `1px solid ${colors.divider}`
+          }}
         >
-          Commit
-        </Button>
-        <HoverToolbar
-          state={this.state.state}
-          onChange={this.onChange.bind(this)}
-        />
-        <Toolbar
-          state={this.state.state}
-          onChange={this.onChange.bind(this)}
-        />
+          <span
+            style={{
+              display: 'block',
+              width: '40px'
+            }}
+          >
+            <R />
+          </span>
+          <Interaction.H1>Haku</Interaction.H1>
+          <Button
+            primary
+            onClick={() => this.props.onCommit(state)}
+          >
+            <FormatBold />
+          </Button>
+        </Container>
+        <HoverToolbar state={state}>
+          <MarkMenu
+            marks={marks}
+            state={state}
+            onChange={this.onChange}
+          />
+          <LinkMenu
+            state={state}
+            onChange={this.onChange}
+          />
+        </HoverToolbar>
+        <SideToolbar state={state}>
+          <BlockMenu
+            blocks={blocks}
+            state={state}
+            onChange={this.onChange}
+          />
+        </SideToolbar>
         <NarrowContainer>
           <SlateEditor
             schema={Schema}
-            onKeyDown={this.onKeyDown}
-            state={this.state.state}
-            onChange={this.onChange.bind(this)}
+            state={state}
+            onChange={this.onChange}
           />
         </NarrowContainer>
       </div>
