+++ lib/editor/schema/index.js
@@ -9,6 +9,8 @@ import {
   Blockquote
 } from './blocks'
 
+import { A } from './inlines'
+
 import {
   Bold,
   Italic,
@@ -38,6 +40,9 @@ export default {
     Span,
     Label,
     Blockquote,
+
+    // Inlines
+    A,
 
     // Marks
     Bold,
