+++ lib/editor/schema/inlines/index.js
@@ -1,0 +1,18 @@
+import createRule, {
+  DEFAULT_PROPS
+} from '../utils/createRule'
+import * as Styleguide from '@project-r/styleguide'
+
+export const A = createRule(
+  'inline',
+  'a',
+  ({ node, children }) => {
+    const { data, attributes } = node
+    return (
+      <Styleguide.A {...attributes} href={data.get('href')}>
+        {children}
+      </Styleguide.A>
+    )
+  },
+  DEFAULT_PROPS.concat('node')
+)
