+++ lib/editor/schema/utils/createRule.js
@@ -18,7 +18,7 @@ export default (
   nodeKind,
   nodeType,
   Component,
-  filterProps = ['style', 'className', 'children', 'id']
+  filterProps = DEFAULT_PROPS
 ) => ({
   match: ({ kind, type }) =>
     kind === nodeKind && type === nodeType,
